feat: complete hw 01-04

+++ src/features/posts/repositories/posts.repository.ts
@@ -2,14 +2,37 @@ import { PostInputDto } from '../dto/postInputDto';
 import { Post } from '../types';
 import { postsCollection } from '../../../db/mongo.db';
 import { ObjectId, WithId } from 'mongodb';
+import { RepositoryNotFoundError } from '../../../core/errors/repository-not-found.error';
+import { PostQueryInput } from '../routes/input/post-query.input';
 
 export const postsRepository = {
-  async findAll(): Promise<WithId<Post>[]> {
-    return postsCollection.find().toArray();
+  async findAll(
+    queryDto: PostQueryInput,
+  ): Promise<{ items: WithId<Post>[]; totalCount: number }> {
+    const { pageSize, pageNumber, sortBy, sortDirection } = queryDto;
+
+    const skip = (pageNumber - 1) * pageSize;
+
+    const items = await postsCollection
+      .find()
+      .sort({ [sortBy]: sortDirection })
+      .skip(skip)
+      .limit(pageSize)
+      .toArray();
+
+    const totalCount = await postsCollection.countDocuments();
+
+    return { items, totalCount };
   },
 
-  async findById(id: string): Promise<WithId<Post> | null> {
-    return postsCollection.findOne({ _id: new ObjectId(id) });
+  async findById(id: string): Promise<WithId<Post>> {
+    const post = await postsCollection.findOne({ _id: new ObjectId(id) });
+
+    if (!post) {
+      throw new RepositoryNotFoundError('Blog not exist');
+    }
+
+    return post;
   },
 
   async create(newPost: Post): Promise<WithId<Post>> {
@@ -18,6 +41,12 @@ export const postsRepository = {
   },
 
   async update(id: string, dto: PostInputDto): Promise<void> {
+    const post = await postsCollection.findOne({ _id: new ObjectId(id) });
+
+    if (!post) {
+      throw new RepositoryNotFoundError('Post not exist');
+    }
+
     const updateResult = await postsCollection.updateOne(
       { _id: new ObjectId(id) },
       {
@@ -31,6 +60,7 @@ export const postsRepository = {
     );
 
     if (updateResult.matchedCount < 1) {
+      //  TODO: заменить на DomainError если такая проверка есть в дз
       throw new Error('Post not exist');
     }
 
@@ -38,14 +68,43 @@ export const postsRepository = {
   },
 
   async delete(id: string): Promise<void> {
+    const post = await postsCollection.findOne({ _id: new ObjectId(id) });
+
+    if (!post) {
+      throw new RepositoryNotFoundError('Post not exist');
+    }
+
     const deleteResult = await postsCollection.deleteOne({
       _id: new ObjectId(id),
     });
 
     if (deleteResult.deletedCount < 1) {
+      //  TODO: заменить на DomainError если такая проверка есть в дз
       throw new Error('Post not exist');
     }
 
     return;
+  },
+
+  async findPostsByBlog(
+    blogId: string,
+    queryDto: PostQueryInput,
+  ): Promise<{ items: WithId<Post>[]; totalCount: number }> {
+    const { pageSize, pageNumber, sortBy, sortDirection } = queryDto;
+
+    const filter = { blogId: blogId };
+
+    const skip = (pageNumber - 1) * pageSize;
+
+    const items = await postsCollection
+      .find(filter)
+      .sort({ [sortBy]: sortDirection })
+      .skip(skip)
+      .limit(pageSize)
+      .toArray();
+
+    const totalCount = await postsCollection.countDocuments();
+
+    return { items, totalCount };
   },
 };
